Refactor uv ecosystem to reuse Python Version and Requirement classes
This commit makes uv depend on dependabot-python and delegates
Uv::Version and Uv::Requirement to their Python equivalents.
This follows the same pattern used in the conda ecosystem.

Related to: https://github.com/github/dependabot-updates/issues/11524
