+++ uv/dependabot-uv.gemspec
@@ -28,6 +28,7 @@ Gem::Specification.new do |spec|
   spec.files        = []
 
   spec.add_dependency "dependabot-common", Dependabot::VERSION
+  spec.add_dependency "dependabot-python", Dependabot::VERSION
 
   common_gemspec.development_dependencies.each do |dep|
     spec.add_development_dependency dep.name, *dep.requirement.as_list
